set up with uv & .env in subfolder
