+++ twilio-countdown-demo/.template.env
@@ -1,0 +1,4 @@
+TWILIO_ACCOUNT_SID=
+TWILIO_MESSAGING_SERVICE_SID=
+TWILIO_AUTH_TOKEN=
+TWILIO_TEST_NUMBER=18777804236
