+++ twilio-countdown-demo/pyproject.toml
@@ -1,0 +1,8 @@
+[project]
+name = "twilio-countdown-demo"
+version = "0.1.0"
+requires-python = ">=3.13"
+dependencies = [
+    "python-dotenv>=1.2.1",
+    "twilio>=9.10.1",
+]
